scai-backend/requirement.txt
flask-caching
flask
bs4
flask-cors
pymilvus
requests
openai
requests_cache
arxiv
llama_index
rapidfuzz
IPython
llama-index-vector-stores-milvus
llama-index-vector-stores-elasticsearch
llama-index-embeddings-huggingface
llama-index-embeddings-ollama
llama-index-llms-ollama
scikit-learn